ProfileView code cleanup
