+++ App/User/Profile/ProfileView.swift
@@ -130,56 +130,7 @@ struct ProfileView: View {
     }
 }
 
-struct SettingsItemView: View {
-    var icon: String
-    var altIcon: String? = nil
-    var color: Color
-    var title: LocalizedStringKey
-    var rotation: Double = 0
-    
-    #if os(iOS)
-    var arrow: SettingsRowArrow? = nil
-    #else
-    var arrow: SettingsRowArrow? = .action
-    #endif
-    
-    enum SettingsRowArrow {
-        case link
-        case action
-    }
-    
-    var body: some View {
-        HStack(spacing: 15) {
-            SettingsSymbolView(
-                icon: icon,
-                altIcon: altIcon,
-                color: color
-            )
-            
-            Text(title)
-                .foregroundColor(.primary)
-            
-            if let arrow {
-                Spacer()
-                Group {
-                    switch arrow {
-                    case .link:
-                        Image(systemName: "arrow.up.forward")
-                    case .action:
-                        Image(systemName: "chevron.forward")
-                    }
-                }
-                .font(.footnote.weight(.bold))
-                .foregroundStyle(.tertiary)
-                .tint(.primary)
-            }
-        }
-        .contentShape(Rectangle())
-    }
-}
-
 struct ProfileView_Previews: PreviewProvider {
-    
     static var previews: some View {
         ProfileView()
     }
